requirements.txt
anndata==0.7.8
igraph==0.9.11
joblib==1.1.1
louvain==0.7.2
matplotlib==3.5.2
munkres==1.1.4
networkx==2.5.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
scanpy==1.9.1
scikit-learn==0.24.2
scikit_misc==0.1.3
scipy==1.5.4
seaborn==0.11.2
spectral==0.23.1
tqdm==4.64.1